rename utils and include hasNoZero
